refactor: UserEntityFixture -> TestUser 로 이름 변경
