+++ src/test/java/com/airjnc/user/service/AuthServiceTest.java
@@ -12,7 +12,7 @@ import com.airjnc.user.util.UserModelMapper;
 import com.testutil.annotation.UnitTest;
 import com.testutil.fixture.AuthLogInReqFixture;
 import com.testutil.fixture.UserRespFixture;
-import com.testutil.fixture.UserEntityFixture;
+import com.testutil.testdata.TestUser;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
@@ -39,7 +39,7 @@ class AuthServiceTest {
   void userShouldBeLoggedIn() {
     //given
     AuthLogInReq authLogInReq = AuthLogInReqFixture.getBuilder().build();
-    UserEntity userEntity = UserEntityFixture.getBuilder().deletedAt(null).build();
+    UserEntity userEntity = TestUser.getBuilder().deletedAt(null).build();
     UserResp userResp = UserRespFixture.getBuilder().build();
     given(userRepository.findWithDeletedByEmail(authLogInReq.getEmail())).willReturn(userEntity);
     given(userModelMapper.userEntityToUserResp(userEntity)).willReturn(userResp);
