+++ src/test/java/com/testutil/config/TestDatabaseConfig.java
@@ -2,7 +2,7 @@ package com.testutil.config;
 
 import com.airjnc.common.config.MybatisConfig;
 import com.airjnc.user.domain.UserEntity;
-import com.testutil.fixture.UserEntityFixture;
+import com.testutil.testdata.TestUser;
 import javax.sql.DataSource;
 import org.mybatis.spring.boot.test.autoconfigure.AutoConfigureMybatis;
 import org.springframework.beans.factory.annotation.Value;
@@ -41,7 +41,7 @@ public class TestDatabaseConfig {
 
   private void initData(DriverManagerDataSource dataSource) {
     JdbcTemplate template = new JdbcTemplate(dataSource);
-    UserEntity user = UserEntityFixture.getBuilder().build();
+    UserEntity user = TestUser.getBuilder().build();
     template.update(
         "insert into `user` (id, email, password, name, gender, phone_number, address, is_active," +
             "birthdate, created_at, updated_at, deleted_at)" +
